Changed Func1 and TestFunc1

+++ code_coverage_4/src/main/java/targil3/Func2.java
@@ -2,7 +2,7 @@ package targil3;
 
 public class Func2 {
 
-    public static String func2 (int p, int q) {
+    public static String func2 (float p, float q) {
 
         if (p != 0) {
             p = p + 10;
@@ -11,7 +11,7 @@ public class Func2 {
         // תקלה: חלוקה ב־0 כאשר p == 0
         q = q / p;
 
-        return String.format("p=%d, q=%d", p, q);
+        return String.format("p=%.2f, q=%.2f", p, q);
     }// end of func2
 
 }//end of class
